Speed up execution 20 to 40%
Added GOTO cache which is updated and used while Basic program is running. This is a simple lookup of target locations of GOTO (and GOSUB) statements. Finding entry in cache means no need to search Basic program text from the beginning, speeding up a lot, depending on how many GOTO/GOSUBs there are, and where they are in program code. Cache is one-way, 32-entries.

Also found a bug in the Basic code, it finds some non-prime numbers too :-)