Adding microcode debug display
Refactored the VGA logic to introduce 3 hardware "windows":

input area (maps to 0x0000 of Basic memory)
basic program (maps to 0x0100 of Basic memory)
microcode symbols - scrolling list of microcode, bottom line is the currently executed instruction. Only displayed in slower CPU speeds.

Symbols file is by ingesting a COE - this is Xilinx (ahem, AMD) specific so will need to be changed to be vendor agnostic. Already a hack was made to MCC compiler to produce this COE.